RUBY-18: Fix various Rubocop warnings and declare dotenv as a dev dependency
