+++ bigcommerce.gemspec
@@ -21,6 +21,7 @@ Gem::Specification.new do |s|
 
   s.add_development_dependency 'bundler'
   s.add_development_dependency 'bundler-audit', '~> 0.6'
+  s.add_development_dependency 'dotenv', '>= 2.0'
   s.add_development_dependency 'pry', '>= 0.12'
   s.add_development_dependency 'rake', '>= 12.0'
   s.add_development_dependency 'rspec', '~> 3.8'
