+++ bin/console
@@ -1,0 +1,6 @@
+#!/usr/bin/env ruby
+
+require 'bundler/setup'
+require 'bigcommerce'
+require 'pry'
+Pry.start
