+++ lib/bigcommerce.rb
@@ -2,22 +2,28 @@
 
 require 'hashie'
 require 'faraday_middleware'
-require 'bigcommerce/version'
-require 'bigcommerce/config'
-require 'bigcommerce/connection'
-require 'bigcommerce/middleware/auth'
-require 'bigcommerce/middleware/http_exception'
-require 'bigcommerce/resources/resource'
+require_relative 'bigcommerce/version'
+require_relative 'bigcommerce/config'
+require_relative 'bigcommerce/connection'
+require_relative 'bigcommerce/path_builder'
+require_relative 'bigcommerce/middleware/auth'
+require_relative 'bigcommerce/middleware/http_exception'
+require_relative 'bigcommerce/resources/resource'
 
 module Bigcommerce
   resources = File.join(File.dirname(__FILE__), 'bigcommerce', 'resources', '**', '*.rb')
-  Dir.glob(resources, &method(:require))
+  Dir.glob(resources).sort.each { |r| require r }
 
   class << self
-    attr_reader :api, :config
+    # @!attribute [r] api
+    #   @return [::Faraday::Connection]
+    attr_reader :api
+    # @!attribute [r] config
+    #   @return [::Bigcommerce::Config]
+    attr_reader :config
 
-    def configure
-      @config = Bigcommerce::Config.new.tap { |h| yield(h) }
+    def configure(&block)
+      @config = Bigcommerce::Config.new.tap(&block)
       @api = Bigcommerce::Connection.build(@config)
     end
   end
