+++ lib/bigcommerce/subresource_actions.rb
@@ -13,31 +13,37 @@ module Bigcommerce
     module ClassMethods
       def all(parent_id, params = {})
         raise ArgumentError if parent_id.nil?
+
         get path.build(parent_id), params
       end
 
       def find(parent_id, resource_id, params = {})
         raise ArgumentError if [parent_id, resource_id].any?(&:nil?)
+
         get path.build([parent_id, resource_id]), params
       end
 
       def create(parent_id, params = {})
         raise ArgumentError if parent_id.nil?
+
         post path.build(parent_id), params
       end
 
       def update(parent_id, resource_id, params = {})
         raise ArgumentError if [parent_id, resource_id].any?(&:nil?)
+
         put path.build([parent_id, resource_id]), params
       end
 
       def destroy(parent_id, resource_id, params = {})
         raise ArgumentError if [parent_id, resource_id].any?(&:nil?)
+
         delete path.build([parent_id, resource_id]), params
       end
 
       def destroy_all(parent_id, params = {})
         raise ArgumentError if parent_id.nil?
+
         delete path.build(parent_id), params
       end
     end
